Add alias and function for executing t twitter client cmds
The t twitter client relies on a certain ruby version, so it needs to be
executed with a specific rvm ruby and gemset, and it's easiest to do
that by just executing in a certain directory. The te function will
execute a t command in that directory in a subshell.